added web abstractions
