+++ lib/definitions.dart
@@ -2,4 +2,4 @@ export 'package:sqlite_async/src/update_notification.dart';
 export 'package:sqlite_async/src/sqlite_connection.dart';
 export 'package:sqlite_async/src/sqlite_queries.dart';
 export 'package:sqlite_async/src/sqlite_open_factory.dart';
-export 'src/sqlite_options.dart';
+export 'package:sqlite_async/src/sqlite_options.dart';
