+++ lib/sqlite_async.dart
@@ -3,8 +3,6 @@
 /// See [SqliteDatabase] as a starting point.
 library;
 
-export 'src/default_sqlite_open_factory.dart';
-export 'src/isolate_connection_factory.dart';
 export 'src/sqlite_connection.dart';
 export 'src/sqlite_database.dart';
 export 'src/sqlite_migrations.dart';
