+++ lib/src/database/native/connection_pool.dart
@@ -2,13 +2,13 @@ import 'dart:async';
 
 import 'package:sqlite3/sqlite3.dart';
 
-import 'mutex.dart';
+import '../../mutex.dart';
+import '../../sqlite_connection.dart';
+import '../../sqlite_open_factory.dart';
+import '../../sqlite_queries.dart';
+import '../../update_notification.dart';
 import 'port_channel.dart';
-import 'sqlite_connection.dart';
 import 'sqlite_connection_impl.dart';
-import 'sqlite_open_factory.dart';
-import 'sqlite_queries.dart';
-import 'update_notification.dart';
 
 /// A connection pool with a single write connection and multiple read connections.
 class SqliteConnectionPool with SqliteQueries implements SqliteConnection {
