+++ lib/src/database/web/web_sqlite_database.dart
@@ -1,0 +1,26 @@
+import 'package:sqlite_async/src/sqlite_connection.dart';
+
+import '../abstract_sqlite_database.dart';
+
+class SqliteDatabase extends AbstractSqliteDatabase {
+  @override
+  bool get closed => throw UnimplementedError();
+
+  @override
+  Future<T> readLock<T>(Future<T> Function(SqliteReadContext tx) callback,
+      {Duration? lockTimeout, String? debugContext}) {
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<T> writeLock<T>(Future<T> Function(SqliteWriteContext tx) callback,
+      {Duration? lockTimeout, String? debugContext}) {
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<void> close() {
+    // TODO: implement close
+    throw UnimplementedError();
+  }
+}
