+++ lib/src/sqlite_options.dart
@@ -1,3 +1,5 @@
+import 'package:sqlite3/wasm.dart';
+
 class SqliteOptions {
   /// SQLite journal mode. Defaults to [SqliteJournalMode.wal].
   final SqliteJournalMode? journalMode;
@@ -11,15 +13,28 @@ class SqliteOptions {
   /// attempt to truncate the file afterwards.
   final int? journalSizeLimit;
 
+  /// The implementation for SQLite
+  /// This is required for Web WASM
+  ///   final wasmSqlite3 =
+  ///       await WasmSqlite3.loadFromUrl(Uri.parse('sqlite3.debug.wasm'));
+  ///   wasmSqlite3.registerVirtualFileSystem(
+  ///       await IndexedDbFileSystem.open(dbName: 'sqlite3-example'),
+  ///       makeDefault: true,
+  ///    );
+  ///  Pass the initialized wasmSqlite3 here
+  final WasmSqlite3? wasmSqlite3;
+
   const SqliteOptions.defaults()
       : journalMode = SqliteJournalMode.wal,
         journalSizeLimit = 6 * 1024 * 1024, // 1.5x the default checkpoint size
-        synchronous = SqliteSynchronous.normal;
+        synchronous = SqliteSynchronous.normal,
+        wasmSqlite3 = null;
 
   const SqliteOptions(
       {this.journalMode = SqliteJournalMode.wal,
       this.journalSizeLimit = 6 * 1024 * 1024,
-      this.synchronous = SqliteSynchronous.normal});
+      this.synchronous = SqliteSynchronous.normal,
+      this.wasmSqlite3 = null});
 }
 
 /// SQLite journal mode. Set on the primary connection.
