+++ lib/utils.dart
@@ -1,0 +1,1 @@
+export 'src/utils.dart';
