+++ test/close_test.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:sqlite_async/sqlite_async.dart';
+import 'package:sqlite_async/src/database/abstract_sqlite_database.dart';
 import 'package:test/test.dart';
 
 import 'util.dart';
@@ -18,7 +19,7 @@ void main() {
       await cleanDb(path: path);
     });
 
-    createTables(SqliteDatabase db) async {
+    createTables(AbstractSqliteDatabase db) async {
       await db.writeTransaction((tx) async {
         await tx.execute(
             'CREATE TABLE test_data(id INTEGER PRIMARY KEY AUTOINCREMENT, description TEXT)');
